backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.11.3
email-validator==2.2.0
cryptography==44.0.2
PyJWT==2.12.1
yfinance==0.2.32
torch==2.6.0